requirements.txt
tqdm~=4.66.1
PyMySQL~=1.1.0
PyYAML~=6.0.1